+++ test/network.js
@@ -3,14 +3,10 @@
 const metatests = require('metatests');
 const metautil = require('..');
 
-const API_EXCHANGE = {
-  host: 'openexchangerates.org',
-  path: '/api/latest.json?app_id=',
-  key: '1f43ea96b1e343fe94333dd2b97a109d',
-};
-
 const getRate = async (currency) => {
-  const { host, path, key } = API_EXCHANGE;
+  const host = 'openexchangerates.org';
+  const path = '/api/latest.json?app_id=';
+  const key = '1f43ea96b1e343fe94333dd2b97a109d';
   const url = `https://${host}/${path}${key}`;
   const data = await metautil.fetch(url);
   const rate = data.rates[currency];
@@ -20,6 +16,14 @@ const getRate = async (currency) => {
 metatests.test('Fetch', async (test) => {
   const rate = await getRate('USD');
   test.strictSame(rate, 1);
+  const method = 'POST';
+  const body = { name: 'Aurelia', age: 43 };
+  const headers = { Authorization: 'Bearer sometoken' };
+  const url = 'https://httpbin.org/anything';
+  const res = await metautil.fetch(url, { headers, body, method });
+  test.strictSame(res.method, method);
+  test.strictSame(res.headers.Authorization, headers.Authorization);
+  test.strictEqual(res.json, body);
   test.end();
 });